[jax2tf] Add support for serialization version 8.
In this version the serialized module contain a StableHLO module
boolean attribute `jax.uses_shape_polymorphism` that specifies
whether the module uses shape polymorphism. If it doesn't then
we do not need to do shape refinement.

Note that we are still keeping the default serialization version to
6, for forward compatibility. However, the serialization unit tests
now run at version 8.

Made Exported.mlir_module a method instead of a propery, to make it
more obvious that it is a derived artifact.