Add a typed `.value` property to JAX config `state` objects.
This change is a follow-on to https://github.com/google/jax/pull/16866, which added a more ABSL-flag API for flags defined with `DEFINE_...`. This change adds a similar typed API for flags defined with `define_..._state`, by adding to the existing `_StateContextManager` object a typed `value` property.

Move a handful of the state objects into the files that use them, and change those files to access the configuration options via the typed `value`. Moving the remainder is left to future PRs.

PiperOrigin-RevId: 551644794