Add backward compatibility test for Pallas Triton serialization.
JAX serializations must be valid for 6 months. Most of this backwards compatibility is provided by StableHLO, but when custom calls are involved we need to ensure that the custom call target exists and can process old serializations.

This change also enables serialization of StableHLO modules that contain calls to `triton_kernel_call`.

PiperOrigin-RevId: 556329711