[pallas] Remove kernel_name and kernel_regeneration_metadata from backend config
This information is now attached as attributes to the MLIR custom_call Op.
When serializing the these additional attributes are attached to the metadata field of the HLO custom_call op.

This is done in order not to interfere with the backward guarantees on the serialization format of custom_call backend_config.

PiperOrigin-RevId: 561771663