[export] Add mlir attributes to the platform index and dimension variables
In presence of multi-platform lowering or shape polymorphism we tag
the platform index and the dimension variable arguments with
MLIR attributions jax.platform_index and jax.dimension_variable.

So far nobody uses these attributes, hence this change should not
change any behavior. This is in preparation for implementing
exporting of effects.