[export] Add tests for multi-platform and cross-platform export
Test for each JAX primitive harness that we can lower it for
multiple platforms and then execute it on multiple platforms
with the same results as the JAX native execution.

This is a large test, covering between 5000-7000 harnesses,
depending on the platform.

Hundreds of harnesses fail this test. In future work we will
address each of the failing harnesses in turn.

PiperOrigin-RevId: 570661115