[export] Adapt lowering caching to work with multi-platform lowering
Previously the mlir.cache_lowering was assuming that a primitive has
a unique lowering in a module for given input and output avals. But with
multi-platform lowering we need to allow multiple lowerings. We
fix this by adding the lowering function to the cache key.

This fixes the multi-platform lowering tests for cumsum and cumprod.