[export] Ensure that we run shape refinement for modules that use multi-platform lowering
For multi-platform lowering we use a constant platform index argument
threaded through all function calls, and we use conditionals
for the lowering of primitives that have multiple lowerings.
In many cases, but not all, these conditionals are removed
by constant folding prior to conversion to HLO, and the XLA
compiler will only see the code for the compilation platform.
However, in some cases these conditionals are not constant-folded
and the XLA compiler will either see code for other platforms
that is does not expect (the TPU tests failing before),
or will simply generate slightly different code
(e.g., the conv_general_dilated tests on CPU,
where we saw numerical differences before).

To address this, we ensure that we run shape refinement
for modules that use multi-platform lowering. The shape refinement
pass already handles inter-procedural constant folding for dimension
value arguments.

At the moment, the platform index argument is modelled as a dimension
value during lowering, so it makes some sense to use the same
shape refinement pass to clean it up before compilation. But
a cleaner solution would be to separate the shape refinement
pass into an interprocedural constant folding, followed by
proper shape refinement. Then we'd introduce a separate
attribute `jax.needs_constant_folding` in addition to
`jax.uses_shape_polymorphism`.

This change fixes the remaining failures in the
multi_platform_export_test for TPU, and the
conv_general_dilated test for CPU.

PiperOrigin-RevId: 571254037