[export] Fix the call_exported in presence of shardings.
Previously, when we call_exported of an Exported module
with shardings, we invoke the right HLO but the enclosing
JAX computation does not know about the shardings of the
called module. This results in errors when invoking the
calling module.

We change call_exported lowering rules to add sharding
constraints for the inputs and the outputs and we add
a check that we call the exported module on the same
number of devices as at export time.