[jax2tf] Add backwards-compatibility fix for native_serialization_platforms.
Prior to https://github.com/google/jax/pull/18104 the
default value of native_serialization_platforms was the
empty tuple, and it seems there were projects, e.g., T5X,
that used that. Now the default value is None, but we
add support for the empty tuple as well to mean the same
as None.

PiperOrigin-RevId: 574133303