[callback] Add a dummy returned values for callbacks without returned values.
Currently, when we call a callback without a returned value we only generate
a Send operation, without a Recv. This means that there is no way to implement ordered effects, where the ordering has to ensure that a previous callback has finished executing on the host before the next one executes. It also seems that there is no way to implement `jax.effects_barrier()` to ensure that the callbacks have all finished executing.

This change adds a dummy returned scalar sent from the host to the device at the end of the callback. This then results in a Recv operation being generated, which fixes the test flakiness we have see in EffectOrderingTest.test_ordered_effect_remains_ordered_across_multiple_devices.

I initially thought that we only need to do this for ordered effects, and issue only Send operations for unordered effects. But then there seems to be no way for `jax.effects_barrier()` to know that all callbacks have finished.

PiperOrigin-RevId: 576398251