Add comb function to jax._src.scipy.special
Here, we add the comb function to jax.scipy.special.
This implementation provides an approximate and efficient
calculation, while also handling the exact case.

As in scipy, we also handle the case where the number of
repetitions is computed.

Finally, we add a test in the tests/lax_scipy_special_functions_test.py
file to maintain testing coverage.

Resolves: #9709