Factor LIBTPU_INIT_ARGS into the compilation cache key.
Workloads that set the environment variable LIBTPU_INIT_ARGS
expect that the cache key will be invalidated if the value
of the variable changes between runs. Today, LIBTPU_INIT_ARGS
is not used in the cache key computation. The fix is to factor
it in similar to what is done with the XLA_FLAGS environment
variable.

Testing: new unit test; test workloads.
PiperOrigin-RevId: 582423420