If parameter to aval is AbstractValue return it.
While exploring the `abstracted_axes` I noticed that when the option
`abstracted_axes` is set, I noticed that passing `AbstractValue`s as inputs
to the `make_jaxpr` will cause it to yield a `TypeError`.

However, passing `AbstractValue`s as inputs to `make_jaxpr` is possible
when `abstracted_axes=None`. To preserve this feature, this is a
possible change.