Do not pass `AbstractValue`s to `get_aval`.
Calling `make_jaxpr` with `AbstractValue`s as parameters is possible
when `abstracted_axes=None`. However, when `abstracted_axes` is set, the
execution path can lead to contexts where `get_aval` is called with
`AbstractValue`s. `get_aval`'s input is expected to be a JAX
value, otherwise a `TypeError` is raised.

To avoid raising a `TypeError`, we need to make sure to never call
`get_aval` with `AbstractValue`s.