[shape_poly] Clean up the shape_poly_test.py
When we recently moved much of shape_poly_test out of jax2tf
we had to add a number of flags to avoid warnings (which are
errors in GitHub CI). Here we clean the tests so that we
can run them without the flags.

The most common problem was that tests were relying on
implicit rank promotion. We added a number of `jnp.expand_dims`
to fix the rank and let the implicit broadcasting do the rest.