Use process_count(backend) in local_devices().
Due to what is arguably a bug, multiple TPU devices in the same job can have the same process index. When determining a process count for, say, CPU, make sure we use the same backend to compute the process_count. Otherwise we might see an apparently out-of-range process index from another backend.

We should perhaps fix the TPU backend not to do this, but that's going to be a bigger change.

PiperOrigin-RevId: 586453157