remove upstream-nightly CI job
Because of C-API changes, this no longer works with upstream numpy without a jaxlib
build, which is too expensive to run as part of a github action. We should replace
with a nightly kokoro job, but in the meantime there's no reason to use these CI
resources for a failing nightly run.