Fix cache_used metric implementation and test.
The cache_used metric is incremented once per task and is
used to determine how many tasks are using the Jax
compilation cache. The current implementation and unit
test are not thread safe. This results in the test
failing when unit tests are executed concurrently.

The fix is to make the implementation thread safe and
to update the test to examine the delta in the metric.

Testing: Cloud TPU VM testing; test workload.
PiperOrigin-RevId: 589174850