Use explicit test case names instead of the numbering system of parameterized
`@parameterized.parameters` interacts very badly with `JaxTestCase.rng`:
the RNG seed is derived from the test method and the test method name
can change if additional test cases are inserted before it. This can cause
CI failures in functions that are completely unrelated to the change that
introduces the breakage.

We should seriously reconsider this strategy. Either all instances of
`parameters` + `self.rng` should be removed or we should find an alternative
strategy for seeding.

PiperOrigin-RevId: 589798050